# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Station API
  description: |
    REST API for Station - Laravel queue management package.

    Authentication is configurable via `station.api.auth`:
    - `token` — requires `Authorization: Bearer <token>` header
    - `none` — no authentication (development only)
  version: 0.1.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  contact:
    name: OJ Baeza
    url: https://github.com/ojbaeza/station
servers:
  - url: '{baseUrl}/api/station'
    description: Station API
    variables:
      baseUrl:
        default: http://localhost
        description: Application base URL
...
paths:
...
  /workers/start:
    post:
      tags: [Workers]
      summary: Start workers for a connection
      operationId: startWorker
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [connection]
              properties:
                connection:
                  type: string
                  description: Queue connection name
                queue:
                  type: string
                  default: default
                  description: Queue name
                workers:
                  type: integer
                  minimum: 1
                  default: 1
                  description: Number of workers to start
      responses:
        '200':
          description: Workers started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /workers/stop:
    post:
      tags: [Workers]
      summary: Stop workers for a connection
      operationId: stopWorker
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [connection]
              properties:
                connection:
                  type: string
                  description: Queue connection name
      responses:
        '200':
          description: Workers stopped
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /workers/stop-external:
    post:
      tags: [Workers]
      summary: Stop an external worker by PID
      operationId: stopExternalWorker
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [pid]
              properties:
                pid:
                  type: integer
                  description: Process ID of the worker to stop
      responses:
        '200':
          description: Worker stopped
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'

  # ── Supervisor ─────────────────────────────────────────────────
...
  /supervisor/start:
    post:
      tags: [Supervisors]
      summary: Start supervisor
      operationId: startSupervisor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [connection]
              properties:
                connection:
                  type: string
                  description: Queue connection name
                queue:
                  type: string
                  default: default
                  description: Queue name
                workers:
                  type: integer
                  minimum: 1
                  default: 1
                  description: Number of workers
      responses:
        '200':
          description: Supervisor started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /queues/pause:
    post:
      tags: [Queues]
      summary: Pause a queue
      operationId: pauseQueue
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [queue]
              properties:
                queue:
                  type: string
                  description: Queue name to pause
                connection:
                  type: string
                  description: Connection name (optional, uses default if omitted)
      responses:
        '200':
          description: Queue paused
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /queues/resume:
    post:
      tags: [Queues]
      summary: Resume a queue
      operationId: resumeQueue
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [queue]
              properties:
                queue:
                  type: string
                  description: Queue name to resume
                connection:
                  type: string
                  description: Connection name (optional, uses default if omitted)
      responses:
        '200':
          description: Queue resumed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
...
components:
...
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
        count:
          type: integer
          description: Number of items affected (when applicable)
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
...
  responses:
...
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: Invalid request parameters
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Bearer token authentication. Set `STATION_API_TOKEN` in your `.env`
        and configure `station.api.auth` to `token`.

    sanctum:
      type: apiKey
      in: cookie
      name: laravel_session
      description: Laravel Sanctum session authentication
security:
  - bearerAuth: []
  - sanctum: []